cloud

+++ ecart-microservices-ws/order-service/src/main/java/com/cts/ecart/dao/OrderDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.cts.ecart.entity.Order;
 
-public interface OrderDao extends JpaRepository<Order, Integer> {
+public interface OrderDao extends JpaRepository<Order, Long> {
 
 }
